build: replace travis with github actions

+++ chain/neutrino.go
@@ -384,7 +384,7 @@ func (s *NeutrinoClient) Rescan(startHash *chainhash.Hash, addrs []btcutil.Addre
 		select {
 		case s.enqueueNotification <- &RescanFinished{
 			Hash:   startHash,
-			Height: int32(bestBlock.Height),
+			Height: bestBlock.Height,
 			Time:   header.Timestamp,
 		}:
 		case <-s.quit:
